refactor(framework): split extension.rs into metadata, marker, inject modules
Pure movement, no behavior change. The mod root keeps discovery
orchestration and the marker-order ABI, re-exporting the same
public surface.
